restaurant_demo/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
SQLAlchemy==2.0.40
Werkzeug==3.1.0
python-dotenv==1.0.1
WTForms==3.2.1
email-validator==2.2.0
psycopg2-binary==2.9.10
pytest==8.3.0
pytest-flask==1.3.0
pytest-cov==5.0.0
